XMP: Automate handling of xmp:ModifyDate and xmp:MetadataDate.
