+++ packages/cli/src/session.ts
@@ -1,0 +1,63 @@
+import { Document, NodeIO, Logger, FileUtils, Transform } from '@gltf-transform/core';
+import { Packet, XMP } from '@gltf-transform/extensions';
+import { formatBytes } from './util';
+
+/** Helper class for managing a CLI command session. */
+export class Session {
+	constructor(private _io: NodeIO, private _logger: Logger, private _input: string, private _output: string) {
+		_io.setLogger(_logger);
+	}
+
+	public static create(io: NodeIO, logger: unknown, input: unknown, output: unknown): Session {
+		return new Session(io, logger as Logger, input as string, output as string);
+	}
+
+	public async transform(...transforms: Transform[]): Promise<void> {
+		const doc = this._input
+			? (await this._io.read(this._input)).setLogger(this._logger)
+			: new Document().setLogger(this._logger);
+
+		// Warn and remove lossy compression, to avoid increasing loss on round trip.
+		for (const extensionName of ['KHR_draco_mesh_compression', 'EXT_meshopt_compression']) {
+			const extension = doc
+				.getRoot()
+				.listExtensionsUsed()
+				.find((extension) => extension.extensionName === extensionName);
+			if (extension) {
+				extension.dispose();
+				this._logger.warn(`Decoded ${extensionName}. Further compression will be lossy.`);
+			}
+		}
+
+		await doc.transform(...transforms, updateMetadata);
+
+		await this._io.write(this._output, doc);
+
+		const { lastReadBytes, lastWriteBytes } = this._io;
+		if (!this._input) {
+			const output = FileUtils.basename(this._output) + '.' + FileUtils.extension(this._output);
+			this._logger.info(`${output} (${formatBytes(lastWriteBytes)})`);
+		} else {
+			const input = FileUtils.basename(this._input) + '.' + FileUtils.extension(this._input);
+			const output = FileUtils.basename(this._output) + '.' + FileUtils.extension(this._output);
+			this._logger.info(
+				`${input} (${formatBytes(lastReadBytes)})` + ` → ${output} (${formatBytes(lastWriteBytes)})`
+			);
+		}
+	}
+}
+
+function updateMetadata(document: Document): void {
+	const root = document.getRoot();
+	const xmpExtension = root.listExtensionsUsed().find((ext) => ext.extensionName === 'KHR_xmp_json_ld') as XMP | null;
+
+	// Do not add KHR_xmp_json_ld to assets that don't already use it.
+	if (!xmpExtension) return;
+
+	const rootPacket = root.getExtension<Packet>('KHR_xmp_json_ld') || xmpExtension.createPacket();
+	const date = new Date().toISOString().substring(0, 10);
+
+	// xmp:MetadataDate should be the same as, or more recent than, xmp:ModifyDate.
+	// https://github.com/adobe/xmp-docs/blob/master/XMPNamespaces/xmp.md
+	rootPacket.setProperty('xmp:ModifyDate', date).setProperty('xmp:MetadataDate', date);
+}
